TravelManagement/src/main/java/com/crossover/project/jsonpojo/MonetaryAmount.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package com.crossover.project.jsonpojo;

import com.crossover.project.entities.Currency;

/**
 *
 * @author waqas.ali2
 */

public class MonetaryAmount {
    private double amount;
    private Currency currency;

    public double getAmount() {
        return amount;
    }

    public void setAmount(double amount) {
        this.amount = amount;
    }


    public Currency getCurrency() {
        return currency;
    }

    public void setCurrency(Currency currency) {
        this.currency = currency;
    }
    
    
}
